Created the user settings screens. linked to backend
Created All frontend part of friends. linked to backend
-Friends screen
-Find Friends screen
-Popup modal for friend requests that updates every time user loads profile screen.

profile screen now correctly shows friends

Calendar linked to the backend

added rest day button for calendar

Added a screen for visiting other user's Profile

Added a screen for viewing Bio in an enlarged format.

Linked Add Workouts to the backend so that it adds new exercises and workouts to db.